% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2 from Lab 9.R
\name{myboot2}
\alias{myboot2}
\title{My Bootstrap Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of resamples done with replacement from the original sample}

\item{x}{the original sample from the population}

\item{fun}{the function applied to each sample to get a statistic}

\item{alpha}{positive difference between 1 and the desired upper quartile of the confidence interval}

\item{cx}{amount by which R will scale text and symbols}

\item{...}{Additional features for the histogram}
}
\value{
a histogram with a point estimate and confidence interval
}
\description{
Uses the bootstrap method to make point and interval estimates
}
\details{
Creates a histogram showing a point estimate and confidence interval from sample
}
\examples{
myboot2(x=rnorm(25,mean=25,sd=10),iter=10000,alpha=0.05)
}
