% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci from Lab 11.R
\name{myci}
\alias{myci}
\title{My confidence interval function for means}
\usage{
myci(x, alpha)
}
\arguments{
\item{x}{a vector containing sample data}

\item{alpha}{positive difference between 1 and the desired confidence level}
}
\value{
a confidence interval for the population mean with the desired confidence level
}
\description{
Calculates the confidence interval for a population mean from a sample
}
\details{
Can calculate the interval for any confidence level
}
\examples{
myci(x=rnorm(30,mean=10,sd=12),alpha=.05)
}
