% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{My function of squares and cubes}
\usage{
myf(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a list of vectors
}
\description{
Produce list of squares and cubes
}
\details{
This is a part of the introduction to package making
}
\examples{
myf(5)
}
