% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik from Lab 10.R
\name{mymaxlik}
\alias{mymaxlik}
\title{My Max Likelihood Function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{a likelihood function}

\item{x}{a vector}

\item{param}{a sequence of values}

\item{...}{}
}
\value{
a graph of the maximum likelihood estimate
}
\description{
Graphs the maximum likelihood estimate for a parameter
}
\details{
Returns the index value of the maximum likelihood estimate and the param and y values for that index value
}
\examples{
mymaxlik(lfun=log(dbinom(x,prob=param,size=20)),x=c(3,3,4,3,4,4,5,4),param=seq(0,1,length=1000))
}
