% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve from Lab 6.R
\name{myncurve}
\alias{myncurve}
\title{My function for plotting normal distributions}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean of the distribution}

\item{sigma}{standard deviation of the distribution}

\item{a}{value for x}
}
\value{
a graph of the normal distribution and the probability of x<a
}
\description{
Plots a normal distribution and shades an area
}
\details{
Calculates the probability of the shaded area on the normal distribution
}
\examples{
myncurve(mu=10,sigma=5,a=6)
}
