% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue from Lab 12.R
\name{mypvalue}
\alias{mypvalue}
\title{My p-value function}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{value of the t statistic}

\item{xmax}{the maximum value of x shown on the graph}

\item{n}{the number of data points in the sample}

\item{alpha}{value that determines the quantiles}
}
\value{
plot of p-value and rejection region for t distribution
}
\description{
Plots the p-value corresponding to a given t-statistic, sample size, and alpha
}
\details{
Plots rejection region for t values as well
}
\examples{
mypvalue=function(t0=2,xmax=5,n=30, alpha=0.05)
}
