% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample from Lab 5.R
\name{mysample}
\alias{mysample}
\title{My sample function}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{size of the sample}

\item{iter}{number of samples taken}

\item{time}{amount of time R suspends its execution of R expressions}
}
\value{
a barplot of each factor's frequency in the sample
}
\description{
Shows frequencies of each factor in a sample
}
\details{
Creates one barplot per iteration showing how often the factors from 1 to 10 appear in a sample of size n
}
\examples{
mysample(n=1000, iter=10, time=1)
}
