% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newmyclt from Lab 8.R
\name{new.myclt}
\alias{new.myclt}
\title{My Central Limit Theorem function}
\usage{
new.myclt(n, iter)
}
\arguments{
\item{n}{size of each sample}

\item{iter}{number of samples taken}
}
\value{
a histogram of the sample means
}
\description{
Plots a histogram of sample means
}
\details{
Makes iter random samples of size n of values between 0 and 5 and plots a histogram of the sample means.
}
\examples{
new.myclt(n=10,iter=10000)
}
